Prevent "nil does not understand keys" error in Ndef completion search
The line `^this.all[Server.default.name].keys` assumes that
every server object has a ProxySpace repository. This isn't true:
the repository is created with the first Ndef for that server.
So your first attempt to use Ndef in a session will produce a
"'keys' not understood" error. This PR guards against this.