Rename Document to BaseDocument
The long-term goal seems to be to have a base class in Sc itself, shared
by the various IDE specific document classes to reuse some of the code.

Until this happens upstream, calling it 'Document' creates collisions
and forces it to be moved to an IDE specific folder. This in itself
causes challenges when you wish to share config between different
environments. For example scnvim and an LSP instance - as scnvim uses
its own IDE identifier/classes.

For now, rename this to BaseDocument, and move it back out of the vscode
specific folder, so it can be readily used in any sclang instance, and
there is no need for a specific config to avoid loading the quark.

Relates to issue https://github.com/scztt/LanguageServer.quark/issues/17